This logic change worked, but it doesn't solve my proplem so I'm just adding the compare flag and 🤷‍♀️ for the short term
