+++ src/common/index.js
@@ -3,5 +3,6 @@ export * from './components/Button'
 export * from './components/Card'
 
 export * from './utils/doubleArray.js'
+export * from './utils/compareCards.js'
 export * from './utils/getShuffledCards.js'
 export * from './utils/shuffleArray.js'
