+++ src/common/utils/index.js
@@ -1,3 +1,4 @@
 export * from './doubleArray.js'
+export * from './compareCards.js'
 export * from './getShuffledCards.js'
 export * from './shuffleArray.js'
